test: cadastrarNovoPedidoDePetComSucesso, buscarPetInexistente and atualizarDadosDePetExistente

+++ src/test/java/com/javarestassuredtemplate/requests/GetPetRequest.java
@@ -4,7 +4,7 @@ import com.javarestassuredtemplate.bases.RequestRestBase;
 import io.restassured.http.Method;
 
 public class GetPetRequest extends RequestRestBase {
-    public GetPetRequest(int petId){
+    public GetPetRequest(long petId){
         requestService = "/pet/"+petId;
         method = Method.GET;
     }
